# Turn ByteIterator into ByteCursor