--- a/src/main/java/io/netty/buffer/api/CompositeBuf.java
+++ b/src/main/java/io/netty/buffer/api/CompositeBuf.java
@@ -293,18 +293,17 @@ final class CompositeBuf extends RcSupport<Buf, CompositeBuf> implements Buf {
 
         // Iterate in reverse to account for src and dest buffer overlap.
         // todo optimise by delegating to constituent buffers.
-        var itr = iterateReverse(srcPos + length - 1, length);
+        var cursor = iterateReverse(srcPos + length - 1, length);
         ByteOrder prevOrder = dest.order();
         // We read longs in BE, in reverse, so they need to be flipped for writing.
         dest.order(ByteOrder.LITTLE_ENDIAN);
         try {
-            while (itr.hasNextLong()) {
-                long val = itr.nextLong();
+            while (cursor.nextLong()) {
                 length -= Long.BYTES;
-                dest.setLong(destPos + length, val);
+                dest.setLong(destPos + length, cursor.getLong());
             }
-            while (itr.hasNextByte()) {
-                dest.setByte(destPos + --length, itr.nextByte());
+            while (cursor.nextByte()) {
+                dest.setByte(destPos + --length, cursor.getByte());
             }
         } finally {
             dest.order(prevOrder);
@@ -312,7 +311,7 @@ final class CompositeBuf extends RcSupport<Buf, CompositeBuf> implements Buf {
     }
 
     @Override
-    public ByteIterator iterate(int fromOffset, int length) {
+    public ByteCursor iterate(int fromOffset, int length) {
         if (fromOffset < 0) {
             throw new IllegalArgumentException("The fromOffset cannot be negative: " + fromOffset + '.');
         }
@@ -326,61 +325,63 @@ final class CompositeBuf extends RcSupport<Buf, CompositeBuf> implements Buf {
         int startBufferIndex = searchOffsets(fromOffset);
         int off = fromOffset - offsets[startBufferIndex];
         Buf startBuf = bufs[startBufferIndex];
-        ByteIterator startIterator = startBuf.iterate(off, Math.min(startBuf.capacity() - off, length));
-        return new ByteIterator() {
+        ByteCursor startCursor = startBuf.iterate(off, Math.min(startBuf.capacity() - off, length));
+        return new ByteCursor() {
             int index = fromOffset;
             final int end = fromOffset + length;
             int bufferIndex = startBufferIndex;
-            ByteIterator itr = startIterator;
+            ByteCursor cursor = startCursor;
+            long longValue = -1;
+            byte byteValue = -1;
 
             @Override
-            public boolean hasNextLong() {
-                return bytesLeft() >= Long.BYTES;
-            }
-
-            @Override
-            public long nextLong() {
-                if (itr.hasNextLong()) {
-                    long val = itr.nextLong();
-                    index += Long.BYTES;
-                    return val;
+            public boolean nextLong() {
+                if (bytesLeft() >= Long.BYTES) {
+                    if (cursor.nextLong()) {
+                        longValue = cursor.getLong();
+                        index += Long.BYTES;
+                    } else {
+                        longValue = nextLongFromBytes(); // Leave index increments to 'nextByte'
+                    }
+                    return true;
                 }
-                if (!hasNextLong()) {
-                    throw new NoSuchElementException();
-                }
-                return nextLongFromBytes(); // Leave index increments to 'nextByte'
+                return false;
             }
 
             private long nextLongFromBytes() {
                 long val = 0;
                 for (int i = 0; i < 8; i++) {
+                    nextByte();
                     val <<= 8;
-                    val |= nextByte();
+                    val |= getByte();
                 }
                 return val;
             }
 
             @Override
-            public boolean hasNextByte() {
-                return index < end;
+            public long getLong() {
+                return longValue;
             }
 
             @Override
-            public byte nextByte() {
-                if (itr.hasNextByte()) {
-                    byte val = itr.nextByte();
+            public boolean nextByte() {
+                if (index < end) {
+                    if (!cursor.nextByte()) {
+                        bufferIndex++;
+                        Buf nextBuf = bufs[bufferIndex];
+                        cursor = nextBuf.iterate(0, Math.min(nextBuf.capacity(), bytesLeft()));
+                        cursor.nextByte();
+                    }
+                    byteValue = cursor.getByte();
                     index++;
-                    return val;
+                    return true;
                 }
-                if (!hasNextByte()) {
-                    throw new NoSuchElementException();
-                }
-                bufferIndex++;
-                Buf nextBuf = bufs[bufferIndex];
-                itr = nextBuf.iterate(0, Math.min(nextBuf.capacity(), bytesLeft()));
-                byte val = itr.nextByte();
-                index++;
-                return val;
+                return false;
+            }
+
+            @Override
+            public byte getByte() {
+                return byteValue;
             }
 
             @Override
@@ -396,7 +397,7 @@ final class CompositeBuf extends RcSupport<Buf, CompositeBuf> implements Buf {
     }
 
     @Override
-    public ByteIterator iterateReverse(int fromOffset, int length) {
+    public ByteCursor iterateReverse(int fromOffset, int length) {
         if (fromOffset < 0) {
             throw new IllegalArgumentException("The fromOffset cannot be negative: " + fromOffset + '.');
         }
@@ -410,60 +411,64 @@ final class CompositeBuf extends RcSupport<Buf, CompositeBuf> implements Buf {
         int startBufferIndex = searchOffsets(fromOffset);
         int off = fromOffset - offsets[startBufferIndex];
         Buf startBuf = bufs[startBufferIndex];
-        ByteIterator startIterator = startBuf.iterateReverse(off, Math.min(off + 1, length));
-        return new ByteIterator() {
+        ByteCursor startCursor = startBuf.iterateReverse(off, Math.min(off + 1, length));
+        return new ByteCursor() {
             int index = fromOffset;
             final int end = fromOffset - length;
             int bufferIndex = startBufferIndex;
-            ByteIterator itr = startIterator;
+            ByteCursor cursor = startCursor;
+            long longValue = -1;
+            byte byteValue = -1;
 
             @Override
-            public boolean hasNextLong() {
-                return bytesLeft() >= Long.BYTES;
-            }
-
-            @Override
-            public long nextLong() {
-                if (itr.hasNextLong()) {
-                    index -= Long.BYTES;
-                    return itr.nextLong();
+            public boolean nextLong() {
+                if (bytesLeft() >= Long.BYTES) {
+                    if (cursor.nextLong()) {
+                        index -= Long.BYTES;
+                        longValue = cursor.getLong();
+                    } else {
+                        longValue = nextLongFromBytes(); // Leave index increments to 'nextByte'
+                    }
+                    return true;
                 }
-                if (!hasNextLong()) {
-                    throw new NoSuchElementException();
-                }
-                return nextLongFromBytes(); // Leave index increments to 'nextByte'
+                return false;
             }
 
             private long nextLongFromBytes() {
                 long val = 0;
                 for (int i = 0; i < 8; i++) {
+                    nextByte();
                     val <<= 8;
-                    val |= nextByte();
+                    val |= getByte();
                 }
                 return val;
             }
 
             @Override
-            public boolean hasNextByte() {
-                return index > end;
+            public long getLong() {
+                return longValue;
             }
 
             @Override
-            public byte nextByte() {
-                if (itr.hasNextByte()) {
-                    byte val = itr.nextByte();
+            public boolean nextByte() {
+                if (index > end) {
+                    if (!cursor.nextByte()) {
+                        bufferIndex--;
+                        Buf nextBuf = bufs[bufferIndex];
+                        int length = Math.min(nextBuf.capacity(), bytesLeft());
+                        cursor = nextBuf.iterateReverse(nextBuf.capacity() - 1, length);
+                        cursor.nextByte();
+                    }
+                    byteValue = cursor.getByte();
                     index--;
-                    return val;
+                    return true;
                 }
-                if (!hasNextByte()) {
-                    throw new NoSuchElementException();
-                }
-                bufferIndex--;
-                Buf nextBuf = bufs[bufferIndex];
-                itr = nextBuf.iterateReverse(nextBuf.capacity() - 1, Math.min(nextBuf.capacity(), bytesLeft()));
-                byte val = itr.nextByte();
-                index--;
-                return val;
+                return false;
+            }
+
+            @Override
+            public byte getByte() {
+                return byteValue;
             }
 
             @Override