… src/main/java/io/netty/buffer/api/memseg/MemSegBuf.java
… src/test/java/io/netty/buffer/api/BufTest.java
… src/test/java/io/netty/buffer/api/benchmarks/ByteIterationBenchmark.java