Motivation:
Cursors are better than iterators in that they only need to check boundary conditions once per iteration, when processed in a loop.
This should make them easier for the compiler to optimise.

Modification:
Change the ByteIterator to a ByteCursor. The API is almost the same, but with a few subtle differences in semantics.
The primary difference is that the cursor movement and boundary condition checking and position movement happen at the same time, and do not need to occur when the values are fetched out of the cursor.
An iterator, on the other hand, needs to throw an exception if "next" is called too many times.

Result:
Simpler code, and faster code as well.

This PR:

```
Benchmark                                    (type)  Mode  Cnt     Score    Error  Units
ByteIterationBenchmark.sum                     heap  avgt   50   478,618 ±  4,147  ns/op
ByteIterationBenchmark.sum                   direct  avgt   50   476,447 ±  6,019  ns/op
ByteIterationBenchmark.sum           composite-heap  avgt   50  2986,791 ± 73,203  ns/op
ByteIterationBenchmark.sum         composite-direct  avgt   50  2670,200 ± 48,309  ns/op
ByteIterationBenchmark.sumReverse              heap  avgt   50   390,934 ±  5,148  ns/op
ByteIterationBenchmark.sumReverse            direct  avgt   50   392,272 ±  6,538  ns/op
ByteIterationBenchmark.sumReverse    composite-heap  avgt   50  2770,515 ± 67,099  ns/op
ByteIterationBenchmark.sumReverse  composite-direct  avgt   50  2680,819 ± 61,209  ns/op
```

And previously, with the iterator pattern:

```
Benchmark                                    (type)  Mode  Cnt     Score    Error  Units
ByteIterationBenchmark.sum                     heap  avgt   50   682,136 ±  3,934  ns/op
ByteIterationBenchmark.sum                   direct  avgt   50   621,579 ±  3,009  ns/op
ByteIterationBenchmark.sum           composite-heap  avgt   50  3469,898 ± 43,624  ns/op
ByteIterationBenchmark.sum         composite-direct  avgt   50  2703,371 ± 53,141  ns/op
ByteIterationBenchmark.sumReverse              heap  avgt   50   438,394 ±  7,681  ns/op
ByteIterationBenchmark.sumReverse            direct  avgt   50   463,282 ±  7,074  ns/op
ByteIterationBenchmark.sumReverse    composite-heap  avgt   50  2713,186 ± 48,320  ns/op
ByteIterationBenchmark.sumReverse  composite-direct  avgt   50  2735,066 ± 91,636  ns/op
```

So reverse iteration is still faster, somehow. Composite got a bit slower, but it ought to be possible to improve on that. There's no good reason, I think, that composite should have a 10x overhead on this.
For non-composite, things are faster overall.