@franz1981 2020-12-02
First look is good, I am a big fan of cursors API vs iteration..tomorrow I will take a better look