@franz1981 2020-12-02
GC apart would be nice to check how the swar indexOf compare vs something using cursors.