@chrisvest 2020-12-07
@normanmaurer I tried to squeeze a bit more performance out of the composite ByteCursor. It helped, but not nearly as much as I had hoped.

Before:

```
Benchmark                                    (type)  Mode  Cnt     Score    Error  Units
ByteIterationBenchmark.sum                     heap  avgt   10   490,014 ±  6,874  ns/op
ByteIterationBenchmark.sum                   direct  avgt   10   489,624 ±  3,916  ns/op
ByteIterationBenchmark.sum           composite-heap  avgt   10  3707,880 ± 82,760  ns/op
ByteIterationBenchmark.sum         composite-direct  avgt   10  3731,500 ± 66,111  ns/op
ByteIterationBenchmark.sumReverse              heap  avgt   10   401,745 ±  3,869  ns/op
ByteIterationBenchmark.sumReverse            direct  avgt   10   398,329 ±  2,268  ns/op
ByteIterationBenchmark.sumReverse    composite-heap  avgt   10  3510,920 ± 80,826  ns/op
ByteIterationBenchmark.sumReverse  composite-direct  avgt   10  3229,770 ± 72,151  ns/op
```

After:

```
Benchmark                                    (type)  Mode  Cnt     Score    Error  Units
ByteIterationBenchmark.sum                     heap  avgt   50   478,618 ±  4,147  ns/op
ByteIterationBenchmark.sum                   direct  avgt   50   476,447 ±  6,019  ns/op
ByteIterationBenchmark.sum           composite-heap  avgt   50  2986,791 ± 73,203  ns/op
ByteIterationBenchmark.sum         composite-direct  avgt   50  2670,200 ± 48,309  ns/op
ByteIterationBenchmark.sumReverse              heap  avgt   50   390,934 ±  5,148  ns/op
ByteIterationBenchmark.sumReverse            direct  avgt   50   392,272 ±  6,538  ns/op
ByteIterationBenchmark.sumReverse    composite-heap  avgt   50  2770,515 ± 67,099  ns/op
ByteIterationBenchmark.sumReverse  composite-direct  avgt   50  2680,819 ± 61,209  ns/op
```