@normanmaurer 2020-12-07
@chrisvest but this before / after is still for iterator vs cursor, correct ?