@chrisvest 2020-12-09
Merging this because all comments have been addressed.